+++ src/main.jsx
@@ -25,6 +25,7 @@ import './v070.css'
 import './v080.css'
 import './v096.css'
 import './v099.css'
+import './v0100.css'
 
 createRoot(document.getElementById('root')).render(
   <React.StrictMode>